requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
python-dotenv
openai
json